+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] / 2022-09-20
+### Features
+- Force to `LoadReferencedAssemblies`
+- Fix `ContainNUnit`
+- Remove `ValidateTestAssemblyNUnitVersion`
+- Add Extension for Debug
+
 ## [1.0.3] / 2022-09-17
 ### Features
 - TestEngine.ContainNUnit()